requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
openai==1.76.2
python-dotenv==1.0.1
sentence-transformers==3.4.1
chromadb==0.5.23
rank-bm25==0.2.2
apscheduler==3.10.4
